CMake (#110)
* Initial CMake

* More CMake

* Start trying to build tests with CMake.

* Fix up tests build and adapt to mw-protocol-3.

* Fix some types for Windows and x86.

* Clean up, and flatten source adding to be simpler.

* Change C++11 to C++14 in CMake.

* Unquote some CMake variables.

* Make CMake search lib/ and include/.

* Make CMake find libmpg123-0.lib and libsndifle-1.lib.

* Make CMake used the mpg123 dll.

* Use FOSS cmake modules for finding libraries.

* Typedef ssize_t as long, to match the libmpg123 Windows header. Needs review.

* Properly rename LIBSNDFILE to SNDFILE in cmake.

* Only define LINK_MPG123_DLL on Windows.

* Add cmake to Travis.

* Change minimum cmake to 2.8.

* Add licenses to contrib cmake files. [no-ci]

* Fix cmake compiler flags.

* Reset libs var.

+++ src/audio/ringbuffer.cpp
@@ -43,12 +43,12 @@ RingBuffer::~RingBuffer()
 	delete[] this->buffer;
 }
 
-unsigned long RingBuffer::WriteCapacity() const
+size_t RingBuffer::WriteCapacity() const
 {
 	return CountCast(PaUtil_GetRingBufferWriteAvailable(this->rb));
 }
 
-unsigned long RingBuffer::ReadCapacity() const
+size_t RingBuffer::ReadCapacity() const
 {
 	return CountCast(PaUtil_GetRingBufferReadAvailable(this->rb));
 }
